refactor(new recipe): rename files
